added check access in methods, corrected delete comment

+++ src/main/java/ru/skypro/homework/service/impl/AdServiceImpl.java
@@ -1,9 +1,9 @@
 package ru.skypro.homework.service.impl;
 
 import lombok.RequiredArgsConstructor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -15,15 +15,18 @@ import ru.skypro.homework.entity.Image;
 import ru.skypro.homework.exception.AdNotFoundException;
 import ru.skypro.homework.mapper.AdsMapper;
 import ru.skypro.homework.repository.AdRepository;
+import ru.skypro.homework.security.MyUserDetails;
 import ru.skypro.homework.service.AdService;
 import ru.skypro.homework.service.ImageService;
 import ru.skypro.homework.service.UserService;
 
 import java.io.IOException;
 import java.util.Collection;
+import java.util.Objects;
 
 import static org.springframework.util.ObjectUtils.isEmpty;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 @Transactional
@@ -33,7 +36,7 @@ public class AdServiceImpl implements AdService {
     private final AdRepository adRepository;
     private final ImageService imageService;
     private final UserService userService;
-    private final Logger log = LoggerFactory.getLogger(AdServiceImpl.class);
+    private final MyUserDetails userDetails;
 
     @Override
     public Collection<AdsDTO> getAllAds(String title) {
@@ -79,11 +82,12 @@ public class AdServiceImpl implements AdService {
     @Override
     public void deleteAd(Long adId) {
         log.info("Request to delete ad by id");
-        if(adRepository.existsById(adId)){
+        Ad ad = adRepository.findById(adId).orElseThrow(AdNotFoundException::new);
+        if(checkAccess(ad)){
             adRepository.deleteById(adId);
         }
         else {
-            throw new AdNotFoundException();
+            throw new RuntimeException("Not access to delete ad");
         }
     }
 
@@ -94,18 +98,19 @@ public class AdServiceImpl implements AdService {
             throw new IllegalArgumentException("Price cannot be negative");
         }
         Ad ad = adRepository.findById(adId).orElseThrow(AdNotFoundException::new);
-
-        adsMapper.updateAds(createAdsDTO,ad);
-        adRepository.save(ad);
+        if (checkAccess(ad)){
+            adsMapper.updateAds(createAdsDTO,ad);
+            adRepository.save(ad);
+        }
 
         return adsMapper.adToAdsDTO(ad);
     }
 
     @Override
-    public Collection<AdsDTO> getUserAllAds(Authentication authentication) {
+    public Collection<AdsDTO> getUserAllAds() {
         log.info("Request to get all user ads");
         Collection<Ad> ads;
-        ads = adRepository.findAllAdsByAuthorId(userService.getAuthorizedUserDto(authentication).getId());
+        ads = adRepository.findAllAdsByAuthorId(userDetails.getIdUserDto());
         return adsMapper.adsToAdsListDto(ads);
     }
 
@@ -113,10 +118,12 @@ public class AdServiceImpl implements AdService {
     public String updateImage(Long adId, MultipartFile image) throws IOException {
         log.info("Request to update image");
         Ad updateAd = adRepository.findById(adId).orElseThrow(AdNotFoundException::new);
-        long idImage = updateAd.getImage().getId();
-        updateAd.setImage(imageService.downloadImage(image));
-        imageService.deleteImage(idImage);
-        adRepository.save(updateAd);
+        if (checkAccess(updateAd)){
+            long idImage = updateAd.getImage().getId();
+            updateAd.setImage(imageService.downloadImage(image));
+            imageService.deleteImage(idImage);
+            adRepository.save(updateAd);
+        }
         return adsMapper.adToAdsDTO(updateAd).getImage();
     }
 
@@ -124,5 +131,15 @@ public class AdServiceImpl implements AdService {
     public byte[] getAdImage(Long adId){
         log.info("Get image of an AD with a ID:" + adId);
         return imageService.getImage(adRepository.findById(adId).orElseThrow(AdNotFoundException::new).getImage().getId());
+    }
+
+    private boolean checkAccess(Ad ad){
+        if(userDetails.getAuthorities().contains(new SimpleGrantedAuthority("ADMIN"))) return true;
+        else{
+            Long adAuthorId = ad.getAuthor().getId();
+            Long userId = Long.valueOf(userDetails.getIdUserDto());
+
+            return Objects.equals(userId, adAuthorId);
+        }
     }
 }
